requirements.txt
# API
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1

# Core dependencies
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-mock==3.11.1
httpx==0.27.0
